phenolab/requirements.txt
# Tested in Python 3.11.2 environment
-e ../
streamlit
python-dotenv
plotly
pandas
polars
openpyxl
boto3